requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
base58==2.1.0
bitarray==1.2.2
blake2b-py==0.1.4
cached-property==1.5.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.5
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.5
eth-bloom==1.0.4
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.2
idna==3.2
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==5.1.0
mypy-extensions==0.4.3
netaddr==0.8.0
parsimonious==0.8.1
plyvel==1.3.0
protobuf==3.18.0
py-ecc==4.1.0
py-evm==0.4.0a4
pycodestyle==2.7.0
pycryptodome==3.10.1
pyethash==0.1.27
pyrsistent==0.18.0
requests==2.26.0
rlp==2.0.1
six==1.16.0
sortedcontainers==2.4.0
toml==0.10.2
toolz==0.11.1
trie==2.0.0a5
typing-extensions==3.10.0.2
urllib3==1.26.6
varint==1.0.2
web3==5.23.1
websockets==9.1
yarl==1.6.3
